add cuda reduce sum kernels

+++ include/storage/table.hpp
@@ -10,7 +10,7 @@ namespace gpu_dbms {
 namespace storage {
 
 // Column-oriented storage for better GPU processing
-using IntColumn = std::vector<int>;
+using IntColumn = std::vector<long>;
 using FloatColumn = std::vector<double>;
 using StringColumn = std::vector<std::string>;
 using BoolColumn = std::vector<bool>;
